requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pydantic-settings==2.1.0
requests==2.31.0
python-dotenv==1.0.0
structlog==23.2.0

# Database dependencies (optional)
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Supabase dependencies
supabase==2.3.0
postgrest==0.13.0
websockets>=11.0.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiosqlite==0.19.0

# Vercel deployment
mangum==0.17.0 